Codes/Backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
mysql-connector-python==8.1.0